feat(#8): rest routes to solve issue using openai and anthropic tool agents ✨
