+++ issue-solver/src/issue_solver/__init__.py
@@ -1,24 +1,21 @@
 import os
 from typing import Callable
-import asyncio
 
-from issue_solver.agents import claude_tools
+from issue_solver.agents import anthropic_tools
 from issue_solver.start_resolution import (
     SolveIssueCommand,
     AgentModel,
     AgentName,
     IssueDescription,
 )
-from issue_solver.agents.openai_with_tools.setup_agent import start_resolution as start_agent_tools_resolution
 
 
 async def main() -> None:
     print("Hello from issue-solver!")
-    await start_agent_tools_resolution()
 
 
 def prepare_solve_issue(
-    generate_patch: Callable[[SolveIssueCommand], None],
+        generate_patch: Callable[[SolveIssueCommand], None],
 ) -> Callable[[], None]:
     return lambda: generate_patch(
         SolveIssueCommand(
@@ -27,10 +24,3 @@ def prepare_solve_issue(
             issue_description=IssueDescription(os.environ["ISSUE_DESCRIPTION"]),
         )
     )
-
-# Add this to properly handle the async main function
-def run_main():
-    asyncio.run(main())
-
-if __name__ == "__main__":
-    run_main()
