--- issue-solver/src/issue_solver/agents/claude_tools/__init__.py
@@ -1,2 +1,0 @@
-def main() -> None:
-    print("Hello from issue-solver claude tools!")
