+++ issue-solver/src/issue_solver/agents/resolution_approaches.py
