+++ issue-solver/src/issue_solver/start_resolution.py
@@ -1,17 +1,17 @@
 from dataclasses import dataclass
-from enum import Enum
+from enum import StrEnum
 
 
-class AgentModel(Enum):
+class AgentModel(StrEnum):
     GPT4O = "gpt-4o"
     GPT4O_MINI = "gpt-4o-mini"
     DEEPSEEK = "deepseek-coder"
-    CLAUDE_35_SONNET = "claude-3.5-sonnet"
-    CLAUDE_35_HAIKU = "claude-3.5-haiku"
+    CLAUDE_35_SONNET = "claude-3-5-sonnet-20241022"
+    CLAUDE_35_HAIKU = "claude-3-5-haiku-20241022"
     QWEN25_CODER = "qwen2.5-coder"
 
 
-class AgentName(Enum):
+class AgentName(StrEnum):
     SWE_AGENT = "swe-agent"
     SWE_CRAFTER = "swe-crafter"
